Selection of increments and histories

+++ feature/counterdetails/src/main/java/com/gurpgork/countthis/feature/counterdetails/CounterDetailsViewModel.kt
@@ -5,19 +5,24 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.gurpgork.countthis.core.designsystem.component.UiMessageManager
 import com.gurpgork.countthis.core.domain.interactors.DeleteCounter
+import com.gurpgork.countthis.core.domain.interactors.DeleteHistory
+import com.gurpgork.countthis.core.domain.interactors.DeleteIncrement
 import com.gurpgork.countthis.core.domain.observers.ObserveCounterDetailed
+import com.gurpgork.countthis.core.extensions.combine
+import com.gurpgork.countthis.core.ui.CounterStateSelector
 import com.gurpgork.countthis.feature.counterdetails.navigation.CounterDetailsArgs
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.collect
-import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 internal class CounterDetailsViewModel @Inject constructor(
     private val deleteCounter: DeleteCounter,
+    private val deleteIncrement: DeleteIncrement,
+    private val deleteHistory: DeleteHistory,
     observeCounterDetailed: ObserveCounterDetailed,
     savedStateHandle: SavedStateHandle,
 //TODO should add sorted increments grouped by date for easy displaying on details page
@@ -27,6 +32,10 @@ internal class CounterDetailsViewModel @Inject constructor(
 
     private val uiMessageManager = UiMessageManager()
 
+    private val incrementSelection = CounterStateSelector()
+    private val historySelection = CounterStateSelector()
+
+    //TODO do I need to have 3 states, one for each tab? as to avoid recompositions when selecting list items?
     private val _state = MutableStateFlow(CounterDetailsViewState())
     val state = _state.asStateFlow()
 
@@ -36,18 +45,92 @@ internal class CounterDetailsViewModel @Inject constructor(
         viewModelScope.launch {
             combine(
                 observeCounterDetailed.flow,
+                incrementSelection.observeSelectedIds(),
+                historySelection.observeSelectedIds(),
+                incrementSelection.observeIsSelectionOpen(),
+                historySelection.observeIsSelectionOpen(),
                 uiMessageManager.message,
-            ){ counter, messages ->
-                CounterDetailsViewState(counter, messages)
-            }.collect{_state.value = it}
+            ) { counter, incrementIds, historyIds, isIncrementSelectionOpen, isHistorySelectionOpen, messages ->
+                CounterDetailsViewState(
+                    counter,
+                    incrementIds,
+                    historyIds,
+                    isIncrementSelectionOpen,
+                    isHistorySelectionOpen,
+                    messages
+                )
+            }.collect { _state.value = it }
         }
     }
 
-    fun deleteCounter(id: Long, listIndex: Int) {
-        viewModelScope.launch {
-            deleteCounter(
-                DeleteCounter.Params(id, listIndex)
-            ).collect()
+
+    fun onLongClick(tab: CounterDetailTabs, id: Long){
+        when(tab){
+            CounterDetailTabs.STATS -> TODO()
+            CounterDetailTabs.DETAILS -> onIncrementLongClick(id)
+            CounterDetailTabs.HISTORY -> onHistoryLongClick(id)
         }
     }
+    fun onClick(tab: CounterDetailTabs, id: Long){
+        when(tab){
+            CounterDetailTabs.STATS -> TODO()
+            CounterDetailTabs.DETAILS -> onIncrementClick(id)
+            CounterDetailTabs.HISTORY -> onHistoryClick(id)
+        }
+    }
+
+    private fun onIncrementLongClick(id: Long) {
+        incrementSelection.onItemLongClick(id)
+    }
+
+    private fun onIncrementClick(id: Long) {
+        incrementSelection.onItemClick(id)
+    }
+
+    private fun onHistoryLongClick(id: Long) {
+        historySelection.onItemLongClick(id)
+    }
+
+    private fun onHistoryClick(id: Long) {
+        historySelection.onItemClick(id)
+    }
+
+    fun deleteCounter(deleteType: DeleteType, id: Long, listIndex: Int) {
+        when (deleteType) {
+            DeleteType.NONE -> assert(false)
+            DeleteType.COUNTER -> {
+                viewModelScope.launch {
+                    deleteCounter(
+                        DeleteCounter.Params(id, listIndex)
+                    ).collect()
+                }
+            }
+
+            DeleteType.INCREMENT -> {
+                viewModelScope.launch {
+                    deleteIncrement(
+                        DeleteIncrement.Params(incrementSelection.getSelectedIds())
+                    ).collect()
+                }
+            }
+
+            DeleteType.HISTORY -> {
+                viewModelScope.launch {
+                    deleteHistory(
+                        DeleteHistory.Params(historySelection.getSelectedIds())
+                    ).collect()
+                }
+            }
+        }
+
+    }
+
+    fun hasSelectedItems(): Boolean {
+        return incrementSelection.getSelectedIds().isNotEmpty() ||
+                historySelection.getSelectedIds().isNotEmpty()
+    }
+}
+
+enum class CounterDetailTabs {
+    STATS, DETAILS, HISTORY
 }
